refactor to PlaneManager. it works
